Move annotation function location
